fix!: factor out Utils module from cli/base
